fix for movies not played as 'mediatype': 'movie'
In player.py Movies played from openmeta sent with correct metadata do not get played as "mediatype = movie" because "infoMeta" is set to false. I couldn't figure out why this was happening but adding a test for "if 'movie' in str(meta):" to set "infoMeta = True" seems to fix this. You can then pass metadata which will be displayed on the OSD.